requirements.txt
# Python 3.10+
Flask>=3.0.0
python-dotenv>=1.0.0
